Move i18n module text key methods to utilities
